pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mcp_financeiro"
version = "0.1.0"
description = "MCPFinanceiro package"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "fastapi",
    "sqlalchemy",
    "python-dotenv",
    "pyjwt",
]

[project.urls]
Homepage = "https://example.com"

